requirements.txt
azure-functions
azure-functions-durable
pandas
dask
pyod
requests
azure-storage-blob 
